requirements.txt
django
djangorestframework
psycopg2-binary
pillow
djangorestframework-simplejwt
drf-yasg
python-decouple
django-filter

django-ckeditor
django-rest-passwordreset